refactor: restructure WebSocket handling and Vue events
This commit primarily refactors the WebSocket handling for the application, moving a lot of the logic and event handling into 'websocket.ts' store, leading to cleaner components. Also included is a modification of Vue events to Vue reactive properties. The updated code implements better separation of concerns, which makes the components more readable and maintainable. The approach also prepares for future additions of features and bug fixes.